+++ packages/uploader/src/react/uploader-provider.tsx
@@ -1,7 +1,7 @@
 import { Box } from '@mui/material';
 import { ReactNode, createContext, lazy, useContext, useRef, useMemo } from 'react';
 import { createPortal } from 'react-dom';
-import { Uploader } from './uploader';
+import Uploader from './uploader';
 import { UploaderProps } from '../types';
 
 interface UploaderProviderProps extends UploaderProps {